refactoring email notification logic
